Move code-search-results to more fitting location
